bot/requirements.txt
python-telegram-bot==22.6
anthropic>=0.78.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
